+++ ios/camerawesome/Sources/camerawesome/Controllers/PhysicalButtons/PhysicalButtonController.m
@@ -11,22 +11,16 @@
 
 - (instancetype)init {
   self = [super init];
-  NSLog(@"📸 PhysicalButtonController: Initializing JPSVolumeButtonHandler");
   self.volumeButtonHandler = [JPSVolumeButtonHandler volumeButtonHandlerWithUpBlock:^{
-    NSLog(@"📸 PhysicalButtonController: Volume UP detected");
     [self tryToSendPhysicalEvent:volume_up];
   } downBlock:^{
-    NSLog(@"📸 PhysicalButtonController: Volume DOWN detected");
     [self tryToSendPhysicalEvent:volume_down];
   }];
-  NSLog(@"📸 PhysicalButtonController: Initialized successfully");
   return self;
 }
 
 - (void)tryToSendPhysicalEvent:(PhysicalButton)physicalButton {
-  NSLog(@"📸 PhysicalButtonController: tryToSendPhysicalEvent called");
   if (debounceTimer != nil) {
-    NSLog(@"📸 PhysicalButtonController: Canceling previous debounce timer");
     [debounceTimer invalidate];
     debounceTimer = nil;
   }
@@ -47,18 +41,15 @@
           return;
       }
 
-      NSLog(@"📸 PhysicalButtonController: Sending event to Flutter: %@", physicalButtonString);
       self->_physicalButtonEventSink(physicalButtonString);
     } else {
-      NSLog(@"❌ PhysicalButtonController: Event sink is nil - cannot send event!");
+      NSLog(@"PhysicalButtonController: Event sink is nil - cannot send event");
     }
   }];
 }
 
 - (void)startListening {
-  NSLog(@"📸 PhysicalButtonController: Starting to listen for volume buttons");
   [self.volumeButtonHandler startHandler:YES];
-  NSLog(@"📸 PhysicalButtonController: Now listening for volume button events");
 }
 
 - (void)stopListening {